Hook BigBrother into backend.
BigBrother now pings the backend on five second intervals sending
information defined using flags.  For example, we can start the Elm
app with

    var app = Elm.BigBrother.worker({
        username: 'samantha',
        site: 'lesswrong.com',
        initialMessage: ''
    });

    app.ports.log.subscribe(function(message) {
        console.log(message);
    });

Where the port is capturing error messages from failed pings.

The site hosting this script has to be added to the CORS settings
in the Django app.