Add test suite; break out message.
Began outlining the frontend.  To make things simple, both the
script to include in user-facing pages and the reporting backend
will use the same models.  This will ensure that the latest data
for each user is basically just saved in the backend and returned.

A test suite was added, mostly for help developing.